follow java coding standard
