+++ src/main/java/hachi/ToDo.java
@@ -1,10 +1,10 @@
-package Hachi;
+package hachi;
 
 /**
  * Represents a to-do task that the user needs to complete. This is a subclass of {@link Task}.
  * A to-do task has a description and a completion status, and can be marked as completed or incomplete.
  */
- public class ToDo extends Task{
+public class ToDo extends Task {
 
     /**
      * Creates a new to-do task with the specified description. The task is initially marked as incomplete.
@@ -19,7 +19,7 @@ package Hachi;
      * Creates a new to-do task with the specified description and completion status.
      *
      * @param description the description of the to-do task
-     * @param status the completion status of the task (true for completed, false for incomplete)
+     * @param status      the completion status of the task (true for completed, false for incomplete)
      */
     public ToDo(String description, boolean status) {
         super(description, status);
