+++ src/main/java/hachi/Ui.java
@@ -1,4 +1,4 @@
-package Hachi;
+package hachi;
 
 /**
  * The Ui class handles the user interface and interaction with the user. It is responsible for
@@ -6,7 +6,7 @@ package Hachi;
  * or deleted. The class also provides feedback on errors and provides a visual representation of
  * the application, such as the welcome and exit screens.
  */
- public class Ui {
+public class Ui {
     private static final String separation = "_________________________________________________";
     private static final int TODO = 0;
     private static final int DEADLINE = 1;
@@ -87,29 +87,29 @@ package Hachi;
      */
     public void success(int code) {
         switch (code) {
-            case TODO: // 0
-                System.out.println("🐶 Hachi.Hachi fetched a new task for you!");
-                break;
+        case TODO: // 0
+            System.out.println("🐶 Hachi.Hachi fetched a new task for you!");
+            break;
 
-            case DEADLINE: // 1
-                System.out.println("🐶 Hachi.Hachi noted your deadline carefully!");
-                break;
+        case DEADLINE: // 1
+            System.out.println("🐶 Hachi.Hachi noted your deadline carefully!");
+            break;
 
-            case EVENT: // 2
-                System.out.println("🐶 Hachi.Hachi added your event to the calendar!");
-                break;
+        case EVENT: // 2
+            System.out.println("🐶 Hachi.Hachi added your event to the calendar!");
+            break;
 
-            case MARK: // 3
-                System.out.println("🐶 Hachi.Hachi wags his tail proudly: 'Hachi.Task complete!'");
-                break;
+        case MARK: // 3
+            System.out.println("🐶 Hachi.Hachi wags his tail proudly: 'Hachi.Task complete!'");
+            break;
 
-            case UNMARK: // 4
-                System.out.println("🐶 Hachi.Hachi whines softly: 'Did we bark too soon?'");
-                break;
+        case UNMARK: // 4
+            System.out.println("🐶 Hachi.Hachi whines softly: 'Did we bark too soon?'");
+            break;
 
-            case DELETE: // 5
-                System.out.println("🐶 Hachi.Hachi dug a hole and buried that task. It’s gone!");
-                break;
+        case DELETE: // 5
+            System.out.println("🐶 Hachi.Hachi dug a hole and buried that task. It’s gone!");
+            break;
         }
 
         // after every success, show the total count
@@ -128,42 +128,42 @@ package Hachi;
      */
     public static void failure(int error) {
         switch (error) {
-            case TODO:
-                System.out.println("Woof! You can't just do nothing!" + "\n" + separation);
-                break;
+        case TODO:
+            System.out.println("Woof! You can't just do nothing!" + "\n" + separation);
+            break;
 
-            case DEADLINE:
-                System.out.println("🐶 Hachi.Hachi paws at you: 'I need both a task and a deadline! "
-                        + "Use it like: deadline <task> /by <time>'");
-                printTimeNote();
-                break;
+        case DEADLINE:
+            System.out.println("🐶 Hachi.Hachi paws at you: 'I need both a task and a deadline! "
+                    + "Use it like: deadline <task> /by <time>'");
+            printTimeNote();
+            break;
 
-            case EVENT:
-                System.out.println("🐶 Hachi.Hachi tilts his head: 'I need a task, a start time, and an end time! "
-                        + "Use it like: event <task> /from <start> /to <end>'");
-                printTimeNote();
-                break;
+        case EVENT:
+            System.out.println("🐶 Hachi.Hachi tilts his head: 'I need a task, a start time, and an end time! "
+                    + "Use it like: event <task> /from <start> /to <end>'");
+            printTimeNote();
+            break;
 
-            case MARK:
-                System.out.println("🐶 Hachi.Hachi is confused: 'which task should i mark?'" + "\n" + separation);
-                break;
+        case MARK:
+            System.out.println("🐶 Hachi.Hachi is confused: 'which task should i mark?'" + "\n" + separation);
+            break;
 
-            case UNMARK:
-                System.out.println("🐶 Hachi.Hachi is confused: 'which task should i unmark?'" + "\n" + separation);
-                break;
+        case UNMARK:
+            System.out.println("🐶 Hachi.Hachi is confused: 'which task should i unmark?'" + "\n" + separation);
+            break;
 
-            case DELETE:
-                System.out.println("🐶 Hachi.Hachi is confused: 'which task should i remove?'" + "\n" + separation);
-                break;
+        case DELETE:
+            System.out.println("🐶 Hachi.Hachi is confused: 'which task should i remove?'" + "\n" + separation);
+            break;
 
-            case UNKNOWN:
-                System.out.println("🐶 Hachi.Hachi tilts his head: 'I don’t understand that command.'\n" + separation);
-                break;
+        case UNKNOWN:
+            System.out.println("🐶 Hachi.Hachi tilts his head: 'I don’t understand that command.'\n" + separation);
+            break;
 
-            case MISSING:
-                System.out.println("🐶 Hachi.Hachi sniffed everywhere, but no task found with that number.");
-                System.out.println("Maybe fetch another number?\n" + separation);
-                break;
+        case MISSING:
+            System.out.println("🐶 Hachi.Hachi sniffed everywhere, but no task found with that number.");
+            System.out.println("Maybe fetch another number?\n" + separation);
+            break;
         }
     }
